GH-74: Move current filters into sub namespace
