+++ Source/Lib/Trakt.NET/Modules/TraktCalendarModule.cs
@@ -6,6 +6,7 @@
     using Requests.Calendars.OAuth;
     using Requests.Handler;
     using Requests.Parameters;
+    using Requests.Parameters.OldFilters;
     using Responses;
     using System;
     using System.Threading;
