+++ Source/Lib/Trakt.NET/Requests/Calendars/ACalendarRequest'1.cs
@@ -4,6 +4,7 @@
     using Extensions;
     using Interfaces;
     using Parameters;
+    using Parameters.OldFilters;
     using System;
     using System.Collections.Generic;
