+++ Source/Lib/Trakt.NET/Requests/Interfaces/ISupportsFilter.cs
@@ -1,6 +1,6 @@
 ﻿namespace TraktNet.Requests.Interfaces
 {
-    using Parameters;
+    using Parameters.OldFilters;
 
     internal interface ISupportsFilter
     {
