+++ Source/Lib/Trakt.NET/Requests/Parameters/OldFilters/TraktMovieFilter.cs
@@ -1,18 +1,25 @@
-﻿namespace TraktNet.Requests.Parameters
+﻿namespace TraktNet.Requests.Parameters.OldFilters
 {
     using Modules;
     using System;
-    using System.Collections.Generic;
+    using System.Threading;
     using Utils;
 
     /// <summary>
-    /// Provides additional filter parameters for all <see cref="TraktCalendarModule" /> methods.<para />
+    /// Provides additional filter parameters for some <see cref="TraktMoviesModule" /> methods.<para />
+    /// Supported by <see cref="TraktMoviesModule.GetMostAnticipatedMoviesAsync(TraktExtendedInfo, TraktMovieFilter, TraktPagedParameters, CancellationToken)" />,
+    /// <see cref="TraktMoviesModule.GetMostCollectedMoviesAsync(Enums.TraktTimePeriod, TraktExtendedInfo, TraktMovieFilter, TraktPagedParameters, CancellationToken)" />,
+    /// <see cref="TraktMoviesModule.GetMostPlayedMoviesAsync(Enums.TraktTimePeriod, TraktExtendedInfo, TraktMovieFilter, TraktPagedParameters, CancellationToken)" />,
+    /// <see cref="TraktMoviesModule.GetMostWatchedMoviesAsync(Enums.TraktTimePeriod, TraktExtendedInfo, TraktMovieFilter, TraktPagedParameters, CancellationToken)" />,
+    /// <see cref="TraktMoviesModule.GetPopularMoviesAsync(TraktExtendedInfo, TraktMovieFilter, TraktPagedParameters, CancellationToken)" />,
+    /// <see cref="TraktMoviesModule.GetRecentlyUpdatedMoviesAsync(System.DateTime?, TraktExtendedInfo, TraktPagedParameters, CancellationToken)" /> and
+    /// <see cref="TraktMoviesModule.GetTrendingMoviesAsync(TraktExtendedInfo, TraktMovieFilter, TraktPagedParameters, CancellationToken)" />.<para />
     /// This class has an fluent interface.
     /// <para>See <a href ="http://docs.trakt.apiary.io/#introduction/filters">"Trakt API Doc - Filters"</a> for more information.</para>
     /// </summary>
-    public class TraktCalendarFilter : TraktCommonFilter
+    public class TraktMovieFilter : TraktCommonMovieAndShowFilter
     {
-        /// <summary>Initializes an <see cref="TraktCalendarFilter" /> instance with the given values.</summary>
+        /// <summary>Initializes an <see cref="TraktMovieFilter" /> instance with the given values.</summary>
         /// <param name="query">An optional query string for titles and descriptions.</param>
         /// <param name="startYear">An optional four digit start year for the years parameter.</param>
         /// <param name="endYear">An optional four digit end year for the years parameter.</param>
@@ -21,8 +28,9 @@
         /// <param name="countries">An optional array of two letter country codes.</param>
         /// <param name="runtimes">An optional <see cref="Range{T}" /> instance for minutes.</param>
         /// <param name="ratings">An optional <see cref="Range{T}" /> instance for ratings.</param>
-        /// <exception cref="ArgumentException">Thrown, if the given query string is null or empty.</exception>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <param name="certifications">An optional array of content certificiations.</param>
+        /// <exception cref="System.ArgumentException">Thrown, if the given query string is null or empty.</exception>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if the given <paramref name="startYear" /> value does not have four digits.
         /// Thrown, if the given <paramref name="endYear" /> value does not have four digits.
         /// Thrown, if the begin value of the given runtimes range is below zero or if its end value is below zero or
@@ -32,49 +40,34 @@
         /// Thrown, if the given language codes array contains a language code, which has more or less than two letters.
         /// Thrown, if the given country codes array contains a country code, which has more or less than two letters.
         /// </exception>
-        public TraktCalendarFilter(string query = null, int? startYear = null, int? endYear = null, string[] genres = null,
-                                   string[] languages = null, string[] countries = null, Range<int>? runtimes = null,
-                                   Range<int>? ratings = null)
-            : base(startYear, endYear, genres, languages, countries, runtimes, ratings)
-        {
-            WithQuery(query);
-        }
-
-        /// <summary>Returns the query string parameter value.</summary>
-        public string Query { get; protected set; }
-
-        /// <summary>Returns, whether the query string parameter is set.</summary>
-        public bool HasQuerySet => !string.IsNullOrEmpty(Query);
-
-        /// <summary>Returns, whether any parameters are set.</summary>
-        public override bool HasValues => base.HasValues || HasQuerySet;
+        public TraktMovieFilter(string query = null, int? startYear = null, int? endYear = null, string[] genres = null,
+                                string[] languages = null, string[] countries = null, Range<int>? runtimes = null,
+                                Range<int>? ratings = null, string[] certifications = null)
+            : base(query, startYear, endYear, genres, languages, countries, runtimes, ratings, certifications) { }
 
         /// <summary>Sets the query string parameter value.</summary>
         /// <param name="query">The query string for titles and descriptions.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentException">Thrown, if the given query string is null or empty.</exception>
-        public TraktCalendarFilter WithQuery(string query)
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentException">Thrown, if the given query string is null or empty.</exception>
+        public new TraktMovieFilter WithQuery(string query)
         {
-            if (query != null && query == string.Empty)
-                throw new ArgumentException("query not valid", nameof(query));
-
-            Query = query;
+            base.WithQuery(query);
             return this;
         }
 
         /// <summary>Deletes the current query value.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public TraktCalendarFilter ClearQuery()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearQuery()
         {
-            Query = null;
+            base.ClearQuery();
             return this;
         }
 
         /// <summary>Sets the start year for the years parameter value.</summary>
         /// <param name="startYear">A four digit year.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Thrown, if the given year does not have four digits.</exception>
-        public new TraktCalendarFilter WithStartYear(int startYear)
+        public new TraktMovieFilter WithStartYear(int startYear)
         {
             base.WithStartYear(startYear);
             return this;
@@ -82,9 +75,9 @@
 
         /// <summary>Sets the end year for the years parameter value.</summary>
         /// <param name="endYear">A four digit year.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Thrown, if the given year does not have four digits.</exception>
-        public new TraktCalendarFilter WithEndYear(int endYear)
+        public new TraktMovieFilter WithEndYear(int endYear)
         {
             base.WithEndYear(endYear);
             return this;
@@ -93,33 +86,33 @@
         /// <summary>Sets the start and end year for the years parameter value.</summary>
         /// <param name="startYear">A four digit year.</param>
         /// <param name="endYear">A four digit year.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Thrown, if at least on of the given year values does not have four digits.</exception>
-        public new TraktCalendarFilter WithYears(int startYear, int endYear)
+        public new TraktMovieFilter WithYears(int startYear, int endYear)
         {
             base.WithYears(startYear, endYear);
             return this;
         }
 
         /// <summary>Deletes the current start year of the years parameter.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearStartYear()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearStartYear()
         {
             base.ClearStartYear();
             return this;
         }
 
         /// <summary>Deletes the current end year of the years parameter.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearEndYear()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearEndYear()
         {
             base.ClearEndYear();
             return this;
         }
 
         /// <summary>Deletes the current years parameter.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearYears()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearYears()
         {
             base.ClearYears();
             return this;
@@ -128,8 +121,8 @@
         /// <summary>Adds multiple Trakt genre slugs to the already existing Trakt genre slugs.</summary>
         /// <param name="genre">A Trakt genre slug.</param>
         /// <param name="genres">An optional array of Trakt genre slugs.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter AddGenres(string genre, params string[] genres)
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter AddGenres(string genre, params string[] genres)
         {
             base.AddGenres(genre, genres);
             return this;
@@ -138,16 +131,16 @@
         /// <summary>Sets the Trakt genre slugs parameter and overwrites already existing ones with given Trakt genre slugs.</summary>
         /// <param name="genre">A Trakt genre slug.</param>
         /// <param name="genres">An optional array of Trakt genre slugs.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter WithGenres(string genre, params string[] genres)
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter WithGenres(string genre, params string[] genres)
         {
             base.WithGenres(genre, genres);
             return this;
         }
 
         /// <summary>Deletes the current genre values.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearGenres()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearGenres()
         {
             base.ClearGenres();
             return this;
@@ -156,11 +149,11 @@
         /// <summary>Adds multiple language codes to the already existing language codes.</summary>
         /// <param name="language">A two letter language code.</param>
         /// <param name="languages">An optional array of two letter language codes.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if one the given language codes has more or less than two letters.
         /// </exception>
-        public new TraktCalendarFilter AddLanguages(string language, params string[] languages)
+        public new TraktMovieFilter AddLanguages(string language, params string[] languages)
         {
             base.AddLanguages(language, languages);
             return this;
@@ -169,19 +162,19 @@
         /// <summary>Sets the language codes parameter and overwrites already existing ones with given language codes.</summary>
         /// <param name="language">A two letter language code.</param>
         /// <param name="languages">An optional array of two letter language codes.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if one the given language codes has more or less than two letters.
         /// </exception>
-        public new TraktCalendarFilter WithLanguages(string language, params string[] languages)
+        public new TraktMovieFilter WithLanguages(string language, params string[] languages)
         {
             base.WithLanguages(language, languages);
             return this;
         }
 
         /// <summary>Deletes the current language values.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearLanguages()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearLanguages()
         {
             base.ClearLanguages();
             return this;
@@ -190,11 +183,11 @@
         /// <summary>Adds multiple country codes to the already existing country codes.</summary>
         /// <param name="country">A two letter country code.</param>
         /// <param name="countries">An optional array of two letter country codes.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if one the given country codes has more or less than two letters.
         /// </exception>
-        public new TraktCalendarFilter AddCountries(string country, params string[] countries)
+        public new TraktMovieFilter AddCountries(string country, params string[] countries)
         {
             base.AddCountries(country, countries);
             return this;
@@ -203,19 +196,19 @@
         /// <summary>Sets the country codes parameter and overwrites already existing ones with given country codes.</summary>
         /// <param name="country">A two letter country code.</param>
         /// <param name="countries">An optional array of two letter country codes.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if one the given country codes has more or less than two letters.
         /// </exception>
-        public new TraktCalendarFilter WithCountries(string country, params string[] countries)
+        public new TraktMovieFilter WithCountries(string country, params string[] countries)
         {
             base.WithCountries(country, countries);
             return this;
         }
 
         /// <summary>Deletes the current country values.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearCountries()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearCountries()
         {
             base.ClearCountries();
             return this;
@@ -224,20 +217,20 @@
         /// <summary>Sets the runtimes value parameter and overwrites already exisiting values with the given ones.</summary>
         /// <param name="begin">The begin value of the runtimes range.</param>
         /// <param name="end">The end value of the runtimes range.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if the given begin value is below zero or if the given end value is below zero or
         /// if the given end value is below the given begin value.
         /// </exception>
-        public new TraktCalendarFilter WithRuntimes(int begin, int end)
+        public new TraktMovieFilter WithRuntimes(int begin, int end)
         {
             base.WithRuntimes(begin, end);
             return this;
         }
 
         /// <summary>Deletes the current runtime values.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearRuntimes()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearRuntimes()
         {
             base.ClearRuntimes();
             return this;
@@ -246,47 +239,59 @@
         /// <summary>Sets the ratings value parameter and overwrites already exisiting values with the given ones.</summary>
         /// <param name="begin">The begin value of ratings range.</param>
         /// <param name="end">The end value of the ratings range.</param>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        /// <exception cref="System.ArgumentOutOfRangeException">
         /// Thrown, if the given begin value is below zero or if the given end value is below zero or
         /// if the given end value is below the given begin value or if the given end value is above 100.
         /// </exception>
-        public new TraktCalendarFilter WithRatings(int begin, int end)
+        public new TraktMovieFilter WithRatings(int begin, int end)
         {
             base.WithRatings(begin, end);
             return this;
         }
 
         /// <summary>Deletes the current rating values.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter ClearRatings()
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearRatings()
         {
             base.ClearRatings();
             return this;
         }
 
-        /// <summary>Deletes all filter parameter values.</summary>
-        /// <returns>The current <see cref="TraktCalendarFilter" /> instance.</returns>
-        public new TraktCalendarFilter Clear()
+        /// <summary>Adds multiple content certifications to the already existing content certifications.</summary>
+        /// <param name="certification">A content certification.</param>
+        /// <param name="certifications">An optional array of content certifications.</param>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter AddCertifications(string certification, params string[] certifications)
         {
-            base.Clear();
-            Query = null;
+            base.AddCertifications(certification, certifications);
             return this;
         }
 
-        /// <summary>
-        /// Creates a key-value-pair list of all set parameter-values.
-        /// Each key-value-pair consists of the parameter name as key and its value.
-        /// </summary>
-        /// <returns>A key-value-pair list of all set parameter-values.</returns>
-        public override IDictionary<string, object> GetParameters()
+        /// <summary>Sets the content certifications parameter and overwrites already existing ones with given content certifications.</summary>
+        /// <param name="certification">A content certification.</param>
+        /// <param name="certifications">An optional array of content certifications.</param>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter WithCertifications(string certification, params string[] certifications)
         {
-            var parameters = base.GetParameters();
+            base.WithCertifications(certification, certifications);
+            return this;
+        }
 
-            if (HasQuerySet)
-                parameters.Add("query", Query);
+        /// <summary>Deletes the current certification values.</summary>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter ClearCertifications()
+        {
+            base.ClearCertifications();
+            return this;
+        }
 
-            return parameters;
+        /// <summary>Deletes all filter parameter values.</summary>
+        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
+        public new TraktMovieFilter Clear()
+        {
+            base.Clear();
+            return this;
         }
     }
 }
