+++ Source/Lib/Trakt.NET/Requests/Parameters/OldFilters/TraktSearchFilter.cs
@@ -1,4 +1,4 @@
-﻿namespace TraktNet.Requests.Parameters
+﻿namespace TraktNet.Requests.Parameters.OldFilters
 {
     using Enums;
     using Modules;
@@ -8,21 +8,14 @@
     using Utils;
 
     /// <summary>
-    /// Provides additional filter parameters for some <see cref="TraktShowsModule" /> methods.<para />
-    /// Supported by <see cref="TraktShowsModule.GetMostAnticipatedShowsAsync(TraktExtendedInfo, TraktShowFilter, TraktPagedParameters, CancellationToken)" />,
-    /// <see cref="TraktShowsModule.GetMostCollectedShowsAsync(TraktTimePeriod, TraktExtendedInfo, TraktShowFilter, TraktPagedParameters, CancellationToken)" />,
-    /// <see cref="TraktShowsModule.GetMostPlayedShowsAsync(TraktTimePeriod, TraktExtendedInfo, TraktShowFilter, TraktPagedParameters, CancellationToken)" />,
-    /// <see cref="TraktShowsModule.GetMostWatchedShowsAsync(TraktTimePeriod, TraktExtendedInfo, TraktShowFilter, TraktPagedParameters, CancellationToken)" />,
-    /// <see cref="TraktShowsModule.GetPopularShowsAsync(TraktExtendedInfo, TraktShowFilter, TraktPagedParameters, CancellationToken)" />,
-    /// <see cref="TraktShowsModule.GetRecentlyUpdatedShowsAsync(DateTime?, TraktExtendedInfo, TraktPagedParameters, CancellationToken)" /> and
-    /// <see cref="TraktShowsModule.GetTrendingShowsAsync(TraktExtendedInfo, TraktShowFilter, TraktPagedParameters, CancellationToken)" />.<para />
+    /// Provides additional filter parameters for the new search methods.<para />
+    /// Supported by <see cref="TraktSearchModule.GetTextQueryResultsAsync(TraktSearchResultType, string, TraktSearchField, TraktSearchFilter, TraktExtendedInfo, TraktPagedParameters, CancellationToken)" />.<para />
     /// This class has an fluent interface.
     /// <para>See <a href ="http://docs.trakt.apiary.io/#introduction/filters">"Trakt API Doc - Filters"</a> for more information.</para>
     /// </summary>
-    public class TraktShowFilter : TraktCommonMovieAndShowFilter
+    public class TraktSearchFilter : TraktCommonFilter
     {
-        /// <summary>Initializes an <see cref="TraktShowFilter" /> instance with the given values.</summary>
-        /// <param name="query">An optional query string for titles and descriptions.</param>
+        /// <summary>Initializes an <see cref="TraktSearchFilter" /> instance with the given values.</summary>
         /// <param name="startYear">An optional four digit start year for the years parameter.</param>
         /// <param name="endYear">An optional four digit end year for the years parameter.</param>
         /// <param name="genres">An optional array of Trakt genre slugs.</param>
@@ -33,10 +26,6 @@
         /// <param name="certifications">An optional array of content certificiations.</param>
         /// <param name="networks">An optional aarray of network names.</param>
         /// <param name="states">An optional aarray of show states. See also <seealso cref="TraktShowStatus" />.</param>
-        /// <exception cref="ArgumentException">
-        /// Thrown, if the given query string is null or empty.
-        /// Thrown, if one of the given states is unspecified.
-        /// </exception>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if the given <paramref name="startYear" /> value does not have four digits.
         /// Thrown, if the given <paramref name="endYear" /> value does not have four digits.
@@ -47,15 +36,21 @@
         /// Thrown, if the given language codes array contains a language code, which has more or less than two letters.
         /// Thrown, if the given country codes array contains a country code, which has more or less than two letters.
         /// </exception>
-        public TraktShowFilter(string query = null, int? startYear = null, int? endYear = null, string[] genres = null,
-                               string[] languages = null, string[] countries = null, Range<int>? runtimes = null,
-                               Range<int>? ratings = null, string[] certifications = null, string[] networks = null,
-                               TraktShowStatus[] states = null)
-            : base(query, startYear, endYear, genres, languages, countries, runtimes, ratings, certifications)
+        public TraktSearchFilter(int? startYear = null, int? endYear = null, string[] genres = null, string[] languages = null,
+                                 string[] countries = null, Range<int>? runtimes = null, Range<int>? ratings = null,
+                                 string[] certifications = null, string[] networks = null, TraktShowStatus[] states = null)
+            : base(startYear, endYear, genres, languages, countries, runtimes, ratings)
         {
+            WithCertifications(null, certifications);
             WithNetworks(null, networks);
             WithStates(TraktShowStatus.Unspecified, states);
         }
+
+        /// <summary>Returns the content certifications parameter value.</summary>
+        public string[] Certifications { get; private set; }
+
+        /// <summary>Returns, whether the content certifications parameter is set.</summary>
+        public bool HasCertificationsSet => Certifications != null && Certifications.Length > 0;
 
         /// <summary>Returns the network names parameter value.</summary>
         public string[] Networks { get; private set; }
@@ -70,31 +65,13 @@
         public bool HasStatesSet => States != null && States.Length > 0;
 
         /// <summary>Returns, whether any parameters are set.</summary>
-        public override bool HasValues => base.HasValues || HasNetworksSet || HasStatesSet;
-
-        /// <summary>Sets the query string parameter value.</summary>
-        /// <param name="query">The query string for titles and descriptions.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        /// <exception cref="ArgumentException">Thrown, if the given query string is null or empty.</exception>
-        public new TraktShowFilter WithQuery(string query)
-        {
-            base.WithQuery(query);
-            return this;
-        }
-
-        /// <summary>Deletes the current query value.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearQuery()
-        {
-            base.ClearQuery();
-            return this;
-        }
+        public override bool HasValues => base.HasValues || HasCertificationsSet || HasNetworksSet || HasStatesSet;
 
         /// <summary>Sets the start year for the years parameter value.</summary>
         /// <param name="startYear">A four digit year.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Thrown, if the given year does not have four digits.</exception>
-        public new TraktShowFilter WithStartYear(int startYear)
+        public new TraktSearchFilter WithStartYear(int startYear)
         {
             base.WithStartYear(startYear);
             return this;
@@ -102,9 +79,9 @@
 
         /// <summary>Sets the end year for the years parameter value.</summary>
         /// <param name="endYear">A four digit year.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Thrown, if the given year does not have four digits.</exception>
-        public new TraktShowFilter WithEndYear(int endYear)
+        public new TraktSearchFilter WithEndYear(int endYear)
         {
             base.WithEndYear(endYear);
             return this;
@@ -113,33 +90,33 @@
         /// <summary>Sets the start and end year for the years parameter value.</summary>
         /// <param name="startYear">A four digit year.</param>
         /// <param name="endYear">A four digit year.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">Thrown, if at least on of the given year values does not have four digits.</exception>
-        public new TraktShowFilter WithYears(int startYear, int endYear)
+        public new TraktSearchFilter WithYears(int startYear, int endYear)
         {
             base.WithYears(startYear, endYear);
             return this;
         }
 
         /// <summary>Deletes the current start year of the years parameter.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearStartYear()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearStartYear()
         {
             base.ClearStartYear();
             return this;
         }
 
         /// <summary>Deletes the current end year of the years parameter.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearEndYear()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearEndYear()
         {
             base.ClearEndYear();
             return this;
         }
 
         /// <summary>Deletes the current years parameter.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearYears()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearYears()
         {
             base.ClearYears();
             return this;
@@ -148,8 +125,8 @@
         /// <summary>Adds multiple Trakt genre slugs to the already existing Trakt genre slugs.</summary>
         /// <param name="genre">A Trakt genre slug.</param>
         /// <param name="genres">An optional array of Trakt genre slugs.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter AddGenres(string genre, params string[] genres)
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter AddGenres(string genre, params string[] genres)
         {
             base.AddGenres(genre, genres);
             return this;
@@ -158,16 +135,16 @@
         /// <summary>Sets the Trakt genre slugs parameter and overwrites already existing ones with given Trakt genre slugs.</summary>
         /// <param name="genre">A Trakt genre slug.</param>
         /// <param name="genres">An optional array of Trakt genre slugs.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter WithGenres(string genre, params string[] genres)
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter WithGenres(string genre, params string[] genres)
         {
             base.WithGenres(genre, genres);
             return this;
         }
 
         /// <summary>Deletes the current genre values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearGenres()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearGenres()
         {
             base.ClearGenres();
             return this;
@@ -176,11 +153,11 @@
         /// <summary>Adds multiple language codes to the already existing language codes.</summary>
         /// <param name="language">A two letter language code.</param>
         /// <param name="languages">An optional array of two letter language codes.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if one the given language codes has more or less than two letters.
         /// </exception>
-        public new TraktShowFilter AddLanguages(string language, params string[] languages)
+        public new TraktSearchFilter AddLanguages(string language, params string[] languages)
         {
             base.AddLanguages(language, languages);
             return this;
@@ -189,19 +166,19 @@
         /// <summary>Sets the language codes parameter and overwrites already existing ones with given language codes.</summary>
         /// <param name="language">A two letter language code.</param>
         /// <param name="languages">An optional array of two letter language codes.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if one the given language codes has more or less than two letters.
         /// </exception>
-        public new TraktShowFilter WithLanguages(string language, params string[] languages)
+        public new TraktSearchFilter WithLanguages(string language, params string[] languages)
         {
             base.WithLanguages(language, languages);
             return this;
         }
 
         /// <summary>Deletes the current language values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearLanguages()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearLanguages()
         {
             base.ClearLanguages();
             return this;
@@ -210,11 +187,11 @@
         /// <summary>Adds multiple country codes to the already existing country codes.</summary>
         /// <param name="country">A two letter country code.</param>
         /// <param name="countries">An optional array of two letter country codes.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if one the given country codes has more or less than two letters.
         /// </exception>
-        public new TraktShowFilter AddCountries(string country, params string[] countries)
+        public new TraktSearchFilter AddCountries(string country, params string[] countries)
         {
             base.AddCountries(country, countries);
             return this;
@@ -223,19 +200,19 @@
         /// <summary>Sets the country codes parameter and overwrites already existing ones with given country codes.</summary>
         /// <param name="country">A two letter country code.</param>
         /// <param name="countries">An optional array of two letter country codes.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if one the given country codes has more or less than two letters.
         /// </exception>
-        public new TraktShowFilter WithCountries(string country, params string[] countries)
+        public new TraktSearchFilter WithCountries(string country, params string[] countries)
         {
             base.WithCountries(country, countries);
             return this;
         }
 
         /// <summary>Deletes the current country values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearCountries()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearCountries()
         {
             base.ClearCountries();
             return this;
@@ -244,20 +221,20 @@
         /// <summary>Sets the runtimes value parameter and overwrites already exisiting values with the given ones.</summary>
         /// <param name="begin">The begin value of the runtimes range.</param>
         /// <param name="end">The end value of the runtimes range.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if the given begin value is below zero or if the given end value is below zero or
         /// if the given end value is below the given begin value.
         /// </exception>
-        public new TraktShowFilter WithRuntimes(int begin, int end)
+        public new TraktSearchFilter WithRuntimes(int begin, int end)
         {
             base.WithRuntimes(begin, end);
             return this;
         }
 
         /// <summary>Deletes the current runtime values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearRuntimes()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearRuntimes()
         {
             base.ClearRuntimes();
             return this;
@@ -266,20 +243,20 @@
         /// <summary>Sets the ratings value parameter and overwrites already exisiting values with the given ones.</summary>
         /// <param name="begin">The begin value of ratings range.</param>
         /// <param name="end">The end value of the ratings range.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// Thrown, if the given begin value is below zero or if the given end value is below zero or
         /// if the given end value is below the given begin value or if the given end value is above 100.
         /// </exception>
-        public new TraktShowFilter WithRatings(int begin, int end)
+        public new TraktSearchFilter WithRatings(int begin, int end)
         {
             base.WithRatings(begin, end);
             return this;
         }
 
         /// <summary>Deletes the current rating values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearRatings()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter ClearRatings()
         {
             base.ClearRatings();
             return this;
@@ -288,46 +265,40 @@
         /// <summary>Adds multiple content certifications to the already existing content certifications.</summary>
         /// <param name="certification">A content certification.</param>
         /// <param name="certifications">An optional array of content certifications.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter AddCertifications(string certification, params string[] certifications)
-        {
-            base.AddCertifications(certification, certifications);
-            return this;
-        }
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter AddCertifications(string certification, params string[] certifications)
+            => AddCertifications(true, certification, certifications);
 
         /// <summary>Sets the content certifications parameter and overwrites already existing ones with given content certifications.</summary>
         /// <param name="certification">A content certification.</param>
         /// <param name="certifications">An optional array of content certifications.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter WithCertifications(string certification, params string[] certifications)
-        {
-            base.WithCertifications(certification, certifications);
-            return this;
-        }
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter WithCertifications(string certification, params string[] certifications)
+            => AddCertifications(false, certification, certifications);
 
         /// <summary>Deletes the current certification values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public new TraktShowFilter ClearCertifications()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter ClearCertifications()
         {
-            base.ClearCertifications();
+            Certifications = null;
             return this;
         }
 
         /// <summary>Adds multiple network names to the already existing network names.</summary>
         /// <param name="network">A network name.</param>
         /// <param name="networks">An optional array of network names.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public TraktShowFilter AddNetworks(string network, params string[] networks) => AddNetworks(true, network, networks);
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter AddNetworks(string network, params string[] networks) => AddNetworks(true, network, networks);
 
         /// <summary>Sets the network names parameter and overwrites already existing ones with given network names.</summary>
         /// <param name="network">A network name.</param>
         /// <param name="networks">An optional array of network names.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public TraktShowFilter WithNetworks(string network, params string[] networks) => AddNetworks(false, network, networks);
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter WithNetworks(string network, params string[] networks) => AddNetworks(false, network, networks);
 
         /// <summary>Deletes the current network values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public TraktShowFilter ClearNetworks()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter ClearNetworks()
         {
             Networks = null;
             return this;
@@ -336,30 +307,31 @@
         /// <summary>Adds multiple show states to the already existing show states.</summary>
         /// <param name="status">A show status. See also <seealso cref="TraktShowStatus" />.</param>
         /// <param name="states">An optional array of show states. See also <seealso cref="TraktShowStatus" />.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentException">Thrown, if one the given show states is unspecified.</exception>
-        public TraktShowFilter AddStates(TraktShowStatus status, params TraktShowStatus[] states) => AddStates(true, status, states);
+        public TraktSearchFilter AddStates(TraktShowStatus status, params TraktShowStatus[] states) => AddStates(true, status, states);
 
         /// <summary>Sets the show states parameter and overwrites already existing ones with given show states.</summary>
         /// <param name="status">A show status. See also <seealso cref="TraktShowStatus" />.</param>
         /// <param name="states">An optional array of show states. See also <seealso cref="TraktShowStatus" />.</param>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
         /// <exception cref="ArgumentException">Thrown, if one the given show states is unspecified.</exception>
-        public TraktShowFilter WithStates(TraktShowStatus status, params TraktShowStatus[] states) => AddStates(false, status, states);
+        public TraktSearchFilter WithStates(TraktShowStatus status, params TraktShowStatus[] states) => AddStates(false, status, states);
 
         /// <summary>Deletes the current state values.</summary>
-        /// <returns>The current <see cref="TraktShowFilter" /> instance.</returns>
-        public TraktShowFilter ClearStates()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public TraktSearchFilter ClearStates()
         {
             States = null;
             return this;
         }
 
         /// <summary>Deletes all filter parameter values.</summary>
-        /// <returns>The current <see cref="TraktMovieFilter" /> instance.</returns>
-        public new TraktShowFilter Clear()
+        /// <returns>The current <see cref="TraktSearchFilter" /> instance.</returns>
+        public new TraktSearchFilter Clear()
         {
             base.Clear();
+            Certifications = null;
             Networks = null;
             States = null;
             return this;
@@ -373,6 +345,9 @@
         public override IDictionary<string, object> GetParameters()
         {
             var parameters = base.GetParameters();
+
+            if (HasCertificationsSet)
+                parameters.Add("certifications", string.Join(",", Certifications));
 
             if (HasNetworksSet)
                 parameters.Add("networks", string.Join(",", Networks));
@@ -390,7 +365,33 @@
             return parameters;
         }
 
-        private TraktShowFilter AddNetworks(bool keepExisting, string network, params string[] networks)
+        private TraktSearchFilter AddCertifications(bool keepExisting, string certification, params string[] certifications)
+        {
+            if (string.IsNullOrEmpty(certification) && (certifications == null || certifications.Length <= 0))
+            {
+                if (!keepExisting)
+                    Certifications = null;
+
+                return this;
+            }
+
+            var certificationsList = new List<string>();
+
+            if (keepExisting && Certifications != null && Certifications.Length > 0)
+                certificationsList.AddRange(Certifications);
+
+            if (!string.IsNullOrEmpty(certification))
+                certificationsList.Add(certification);
+
+            if (certifications != null && certifications.Length > 0)
+                certificationsList.AddRange(certifications);
+
+            Certifications = certificationsList.ToArray();
+
+            return this;
+        }
+
+        private TraktSearchFilter AddNetworks(bool keepExisting, string network, params string[] networks)
         {
             if (string.IsNullOrEmpty(network) && (networks == null || networks.Length <= 0))
             {
@@ -416,7 +417,7 @@
             return this;
         }
 
-        private TraktShowFilter AddStates(bool keepExisting, TraktShowStatus status, params TraktShowStatus[] states)
+        private TraktSearchFilter AddStates(bool keepExisting, TraktShowStatus status, params TraktShowStatus[] states)
         {
             if ((status == null || status == TraktShowStatus.Unspecified) && (states == null || states.Length <= 0))
             {
