+++ Source/Lib/Trakt.NET/Requests/Search/SearchTextQueryRequest.cs
@@ -2,7 +2,7 @@
 {
     using Enums;
     using Interfaces;
-    using Parameters;
+    using Parameters.OldFilters;
     using System;
     using System.Collections.Generic;
