+++ Source/Tests/Trakt.NET.Tests/Modules/TraktCalendarModule/TraktCalendarModuleTests_Tests_Json_Data.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
 
     public partial class TraktCalendarModule_Tests
     {
