+++ Source/Tests/Trakt.NET.Tests/Modules/TraktSearchModule/TraktSearchModule_Tests_Json_Data.cs
@@ -2,6 +2,7 @@
 {
     using TraktNet.Enums;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
 
     public partial class TraktSearchModule_Tests
     {
