+++ Source/Tests/Trakt.NET.Tests/Requests/Calendars/OAuth/CalendarUserNewShowsRequest_Tests.cs
@@ -9,6 +9,7 @@
     using TraktNet.Requests.Base;
     using TraktNet.Requests.Calendars.OAuth;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using Xunit;
 
     [Category("Requests.Calendars.OAuth.Shows")]
