+++ Source/Tests/Trakt.NET.Tests/Requests/Movies/AMoviesMostPWCRequest'1_Tests.cs
@@ -9,6 +9,7 @@
     using TraktNet.Requests.Base;
     using TraktNet.Requests.Movies;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using Xunit;
 
     [Category("Requests.Movies.Lists")]
