+++ Source/Tests/Trakt.NET.Tests/Requests/Movies/AMoviesRequest'1_Tests.cs
@@ -8,6 +8,7 @@
     using TraktNet.Requests.Base;
     using TraktNet.Requests.Movies;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using Xunit;
 
     [Category("Requests.Movies.Lists")]
