+++ Source/Tests/Trakt.NET.Tests/Requests/Movies/MoviesTrendingRequest_Tests.cs
@@ -6,6 +6,7 @@
     using Traits;
     using TraktNet.Requests.Movies;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using Xunit;
 
     [Category("Requests.Movies.Lists")]
