+++ Source/Tests/Trakt.NET.Tests/Requests/Shows/AShowsRequest'1_Tests.cs
@@ -7,6 +7,7 @@
     using Traits;
     using TraktNet.Requests.Base;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using TraktNet.Requests.Shows;
     using Xunit;
