+++ Source/Tests/Trakt.NET.Tests/Requests/Shows/ShowsMostAnticipatedRequest_Tests.cs
@@ -5,6 +5,7 @@
     using System.Collections.Generic;
     using Traits;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using TraktNet.Requests.Shows;
     using Xunit;
