+++ Source/Tests/Trakt.NET.Tests/Requests/Shows/ShowsMostWatchedRequest_Tests.cs
@@ -6,6 +6,7 @@
     using Traits;
     using TraktNet.Enums;
     using TraktNet.Requests.Parameters;
+    using TraktNet.Requests.Parameters.OldFilters;
     using TraktNet.Requests.Shows;
     using Xunit;
